contrib: kickstart: make greybox topology starter a service
With this patch, the included kickstart file creates a systemd service
to start the default greybox topology (if an .imn file is present in
the /etc/topgen directory), rather than launch it blindly from rc.local.
This allows us to ensure core-daemon.service is running, so the topology
can be launched successfully.

Signed-off-by: Gabriel Somlo <glsomlo@cert.org>