contrib: kickstart: don't configure BGP
While still installing quagga as a CORE dependency, TopGen (the full
machine OR the container) no longer advertises a default route into
BGP. It is the job of TopGen's upstream router(s), whether operated
as a standalone machine or as a container, to advertise routes with
TopGen as the next-hop into BGP. TopGen still has all the IPs set up
on its loopback, and will respond to traffic sent to them, but no
longer needs to deal with routing and BGP at all. This simplifies
things via a better "division of labor", where Quagga containers
(or external routers) deal with routing, and TopGen deals with
application-level virtualization exclusively.

Signed-off-by: Gabriel Somlo <glsomlo@cert.org>