topgen-mkdns.sh: allow additional source of <ip_addr fqdn> mappings
In addition to Web (hosts.nginx) and Mail (hosts.vmail) sources, allow
an arbitrary user-specified hosts file to be added to the DNS database.

Since /var/lib/topgen/etc/hosts.* are also used to configure service IPs
on the TopGen server's loopback interface, this provides a way to add
records to DNS without also adding service IPs to TopGen. The canonical
use case for this option is point-to-point BGP backbone router interface
ip-to-hostname mappings.

Signed-off-by: Gabriel Somlo <glsomlo@cert.org>