Avoid using IP addresses in the 0.0.0.0/8 block
The scraper needs a fallback IP address for when a vhost can no longer
be resolved in DNS after the wget-based collection phase is completed.

Until now, the made-up address of 0.64.83.76 was used for this role.
However, the 0.0.0.0/8 block is listed as reserved by IANA, and many
tools (traceroute, ping) refuse to work with addresses in that range.

This patch replaces 0.64.83.76 with 1.1.1.1 for the made-up, fallback
address used with unresolvable fqdn host names.

Signed-off-by: Gabriel Somlo <glsomlo@cert.org>