topgen-scrape.sh: don't generate 'default_server' for port 80
Until now, an elegant and concise construct was used to provide
*all* HTTP (port-80) vhosts in a single block:

	server {
		listen 80 default_server;
		server_name \$http_host;
		root $TOPGEN_VHOSTS/\$http_host;
	}

Unfortunately, this clashes with the way most distro packages
for nginx are shipped (with a default http server, also marked
as 'default_server'). As there can only be one 'default_server',
we're faced with having to edit the default nginx.conf, which is
extremely rude if done at install time via the package manager
(e.g., in the %post section of an RPM package), and cumbersome
if left as a TODO item for the users to perform manually after
installing TopGen.

Additionally, due to the impossibility of generating a single
certificate with all our vhost names stuffed in its [alt_names]
section without also exceeding the 16K size limit beyond which
Firefox has been observed to stop accepting the certificate, we
already MUST enumerate dedicated "server {...}" sections for
each vhost to support HTTPS functionality.

This patch modifies topgen-scrape.sh to no longer include the
wildcard HTTP "server {...}" section, and instead add a
"listen 80" statement to each dedicated "server {...}" block,
right next to each corresponding "listen 443 ssl" statement.
This allows for peaceful coexistence with the preinstalled
'default_server' shipping with nginx distribution packages.

Signed-off-by: Gabriel Somlo <glsomlo@cert.org>